Add CreateCompany component to add new companies to our data set
The CreateCompany component consist of a form to submit FE values
to the graphql server via the addCompanyMutation.

Also add links in order to navigate between the create form and
the main list.

For better readability, the graphql queries and mutation were
moved to a separate files using `definition.graphql.js` naming.

`react-router-dom` package was added as well to using Link component